Removed break statement because of strange issues.
In one of my deployments, the break in the Begin block of Send-EmailLog was causing the entire script to exit for some reason. I can't reproduce it outside of that environment, so I can't add a test for it, unfortunately. Replaced it with a  variable instead, which should behave identically.